Update arduino_portenta_h7_m7.overlay
Added in the other two Wire objects to match the MBED version.
i2cs = <&i2c3>, <&i2c1>, <&i2c4>;

I checked it out with simple wire scanner sketch that checks Wire, Wire1 and Wire2.
And then tried adding QWIIC Button first to the I2C0 section of breakout board and
it was found on Wire.

Moved it to I2C1 section and found with Wire1 and then moved to I2C2 section and found on Wire2